app/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
google-cloud-bigquery>=3.14.0
sqlalchemy>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
python-multipart>=0.0.6
redis>=5.0.0
